requirements.txt
scrapy
scrapy-splash
kafka-python
pymongo